stm32: move osc & clock logic to machine package
